studio/server/requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
aiofiles
